Modify the beat macro to accept a list of tasks to schedule
The modification unearthed a few problems that needed to be addressed.

First of all, the CronScheduleError was a standalone error, but it
is more useful to make it part of the BeatError family, so that
users may expect only one type of error when using the beat module.
This is also needed by our macro, that internally creates a function
which should return only one type of error (i.e., BeatError).

Another problem was that the BeatBuilder's build method
produced a CeleryError instead of a BeatError, which was weird.
This was due to the call to configure_task_routes on the broker,
which in turn calls Rule::new.
There is no specific reason why Rule::new should produce a
CeleryError, so it now produces a BadRoutingPattern which is
convertible to BrokerError which is convertible to both
CeleryError and BeatError.