fix sending a message from a rust client to a python worker with redis broker (#371)
* fix sending a message from a rust client to a python worker with redis broker

This adds a delivery info to messages. It basically does the same as
https://github.com/rusty-celery/rusty-celery/pull/253/files but allows for the
broker to customize the delivery info to modify the exchange / routing key wich
can be used from the amqp broker (though this PR does no surface an interface to
it).

In particluar, this fixes the issue of sending a message from a rust client to a
python worker. Without this fix this fails with:

Unrecoverable error: KeyError('exchange')

as kombu/transport/redis.py tries to access the exchange and routing key from
the delivery info.

* fix CI failure

The CI failure of `check-minimal-versions` is caused by this:
- https://github.com/dtolnay/proc-macro2/issues/356
- https://github.com/rust-lang/rust/issues/113152

Not having thiserror pinned at 1.0.2 allow the proc-macro2 crate to be updated to the recent version which fixes this.